+++ CleverPush/Source/CPBannerCardContainer.h
@@ -12,7 +12,7 @@
 - (void)navigateToNextPage:(NSString *)value;
 @end
 
-@interface CPBannerCardContainer : UICollectionViewCell <UITableViewDelegate, UITableViewDataSource>
+@interface CPBannerCardContainer : UICollectionViewCell <UITableViewDelegate, UITableViewDataSource , WKNavigationDelegate, WKUIDelegate, WKScriptMessageHandler>
 @property (weak, nonatomic) IBOutlet UITableView *tblCPBanner;
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *tblCPBannerHeightConstraint;
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *bottomViewBannerConstraint;
@@ -31,6 +31,7 @@
 @property (nonatomic, weak) id <HeightDelegate> delegate;
 @property (nonatomic, weak) id <NavigateNextPage> changePage;
 @property (nonatomic, assign) NSString *voucherCode;
+@property (nonatomic) NSInteger currentScreenIndex;
 
 - (void)setActionCallback:(CPAppBannerActionBlock)callback;
 - (void)setDynamicCloseButton:(BOOL)closeButtonEnabled;