+++ CleverPush/Source/CPUtils.h
@@ -32,5 +32,10 @@
 + (NSUserDefaults *)getUserDefaultsAppGroup;
 + (UIColor *)readableForegroundColorForBackgroundColor:(UIColor*)backgroundColor;
 + (NSString *)replaceString:(NSString *)originalString withReplacement:(NSString *)replacement inString:(NSString *)inputString;
++ (NSString *)cleverPushJavaScript;
++ (NSString *)generateBannerHTMLStringWithFunctions:(NSString *)content;
++ (NSArray<NSString *> *)scriptMessageNames;
++ (void)configureWebView:(WKWebView *)webView;
++ (void)userContentController:(WKUserContentController *)userContentController didReceiveScriptMessage:(WKScriptMessage *)message;
 
 @end