+++ Frameworks/CleverPush.xcframework/ios-arm64/CleverPush.framework/Headers/CPAppBannerCarouselBlock.h
@@ -1,5 +1,6 @@
 #import <Foundation/Foundation.h>
 #import "CPAppBannerBlock.h"
+#import "CPAppBannerBackground.h"
 
 @interface CPAppBannerCarouselBlock : NSObject
 
@@ -7,6 +8,8 @@
 @property (nonatomic, strong) NSString *content;
 @property (nonatomic, strong) NSMutableArray<CPAppBannerBlock*> *blocks;
 @property (assign) BOOL isScreenClicked;
+@property (assign) BOOL isScreenAlreadyShown;
+@property (nonatomic, strong) CPAppBannerBackground *background;
 
 - (id)initWithJson:(NSDictionary*)json;