+++ Frameworks/CleverPush.xcframework/ios-arm64/CleverPush.framework/Headers/CleverPushInstance.h
@@ -167,6 +167,7 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (void)setAppBannerShownCallback:(CPAppBannerShownBlock)callback;
 - (void)setApiEndpoint:(NSString*)apiEndpoint;
 - (void)setAuthorizerToken:(NSString*)authorizerToken;
+- (void)setCustomTopViewController:(UIViewController*)viewController;
 - (void)updateBadge:(UNMutableNotificationContent*)replacementContent API_AVAILABLE(ios(10.0));
 - (void)addStoryView:(CPStoryView*)storyView;
 - (void)updateDeselectFlag:(BOOL)value;
@@ -188,6 +189,7 @@ extern NSString * const CLEVERPUSH_SDK_VERSION;
 - (NSString*)getSubscriptionId;
 - (NSString*)getApiEndpoint;
 - (NSString*)channelId;
+- (UIViewController*)getCustomTopViewController;
 
 - (UIColor*)getBrandingColor;